Fix issues causing console spam
- Don't attempt to translate objects
- Sanitize props passed to react DOM elements
- Remove redundant refresh prop on receivers list view
- Separate <Table> from <SimpleShowLayout> components with a <Fragment>
  to prevent props being passed
- Remove redundant <div> from card views
