+++ Development/src/components/sanitizeRestProps.js
@@ -1,0 +1,25 @@
+import omit from 'lodash/omit';
+
+const sanitizeRestProps = props =>
+    omit(props, [
+        'addLabel',
+        'allowEmpty',
+        'basePath',
+        'cellClassName',
+        'className',
+        'formClassName',
+        'headerClassName',
+        'label',
+        'linkType',
+        'link',
+        'locale',
+        'record',
+        'resource',
+        'sortable',
+        'sortBy',
+        'source',
+        'textAlign',
+        'translateChoice',
+    ]);
+
+export default sanitizeRestProps;
